신규: #13 경기 통계 초기화 기능 구현

+++ common-service/src/main/java/com/kickoff/service/common/application/constant/ApplicationConfigConstants.java
@@ -3,14 +3,14 @@ package com.kickoff.service.common.application.constant;
 import java.util.List;
 
 public class ApplicationConfigConstants {
-  public static final List<Integer> AVAILABLE_LEAGUE_API_FOOTBALL_LEAGUE_IDS = List.of(
-    //    2  // UEFA
-    39  // EPL
-    , 140  // LaLiga
-    , 135  // Serie A
-    , 78 // Bundesliga
-    , 61 // France Ligue 1
-    , 292 // K League 1
+  public static final List<Long> AVAILABLE_LEAGUE_API_FOOTBALL_LEAGUE_IDS = List.of(
+    //    2L  // UEFA
+    39L  // EPL
+    , 140L  // LaLiga
+    , 135L  // Serie A
+    , 78L // Bundesliga
+    , 61L // France Ligue 1
+    , 292L // K League 1
   );
 
   private ApplicationConfigConstants() {
